Task fetching and notes adding
Added the ability to fetch tasks on the tasks page as well as to add notes to and  task and then fetch all notes. As part of this a notes form and viewer were added. Additionally profile fetching now fetches the logged in user as opposed to just the first user